Disable yoda
yoda condition was forbidden before, but I found we shouldn't.
When checking if the number is inside range as follows, it is forbidden to place number at the left of the variable:

if (200 <= res.status && res.status < 300) {}

In above example, 200 can't be placed in the left of res.status. You need to write as follows:

if (res.status => 200 && res.status < 300) {}